requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
pytest==7.4.3
pytest-django==4.7.0
pytest-factoryboy==2.5.1
dj-config-url~=0.1.1